[nasa/nos3#97] Streamlined FSW device file and resolved conversion issue in simulator;

+++ sim/inc/generic_css_data_point.hpp
@@ -17,7 +17,7 @@ namespace Nos3
         /* Accessors */
         /* Provide the hardware model a way to get the specific data out of the data point */
         std::string to_string(void) const;
-        std::vector<double> getValues(void) const {return _generic_css_data;}
+        std::vector<float> getValues(void) const {return _generic_css_data;}
     
     private:
         /* Disallow these */
@@ -27,7 +27,7 @@ namespace Nos3
 
         /* Specific data you need to get from the data provider to the hardware model */
         /* You only get to this data through the accessors above */
-        std::vector<double> _generic_css_data;
+        std::vector<float> _generic_css_data;
         static const int numChannels = 6;
     };
 }
